vector_requirements.txt
# Vector Database Requirements
# Install on desktop/laptop (NOT Android/Termux)

# Core dependencies
torch>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Utilities
numpy>=1.24.0
scikit-learn>=1.3.0

# Progress bars
tqdm>=4.65.0

# Already installed from main project
# sqlite3 (built-in)
# requests (for potential model downloads)
